rf HomeWork Task7, ad HomeWork Task8

+++ HomeWorks/Task7/src/com/company/Bike.java
@@ -8,7 +8,7 @@ public class Bike extends Transport {
     private double oilAmount;
 
     public Bike(String model, double fuelConsumption,
-                int parameter,
+                double parameter,
                 double oilConsuption) {
         super(model, fuelConsumption,parameter);
         this.oilConsuption = oilConsuption;
